+++ CHANGELOG.md
@@ -18,6 +18,14 @@
   - The loading configuration includes an optional trajectory preset provider `TrajectoryHierarchyPresetProvider`
 - Remove `alignMotif` methods (& pecos-integration) as the strucmotif service now reports RMSD and transformations for all hits
 
+## [1.9.3] - 2021-10-18
+### Bug fixes
+- Rewire index parameter of symmetry preset
+
+## [1.9.2] - 2021-10-15
+### Bug fixes
+- removed implicit 'chain-mode', must be explicitly requested by setting 'extendToChain' to true as part of the part
+
 ## [1.9.1] - 2021-10-14
 ### Bug fixes
 - More robust 'chain-mode' check in feature preset